palitra_task/requirements.txt
django
djangorestframework
django-filter
pillow
djongo
django-redis
djangorestframework-simplejwt
drf-yasg